audio: organize network frame support
